refactor: #229 admin categories God Component 분리
SortableCategoryRow를 인라인 정의에서 src/components/admin/SortableCategoryRow.tsx로 추출.
